Modificación de las tablas y sus relaciones. Cambios en los estilos.

+++ app/Http/Controllers/AccountController.php
@@ -4,6 +4,8 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use App\User;
+use App\Account;
+use Illuminate\Support\Str;
 use App\Game;
 use App\Category;
 use Auth;
@@ -12,51 +14,38 @@ use Illuminate\Support\Facades\Storage;
 class AccountController extends Controller
 {
   /**
-   * Display the specified resource.
+   * Display the specified user.
    *
-   * @param  string  $name
+   * @param  string  $slug
    * @return \Illuminate\Http\Response
    */
-  public function show($name)
+  public function show($slug)
   {
-    $user = User::where("name", $name)->first();
-    $games = Game::where("user_id", $user->id)->get();
-
-    /* Guarda el nombre de  las categorías afectadas y elimina lass repetidas */
-    $categories_name = [];
-    foreach ($games as $game) {
-      array_push($categories_name, $game->category->name);
-    }
-    $results = array_unique($categories_name);
-
-    /* Guardo las categorías cuyo nombre coincida con las del array */
-    $categories = [];
-    foreach ($results as $result) {
-      array_push($categories, Category::where( "name", $result)->first() );
-    }
+    $account = Account::where("slug", $slug)->first();
+    $user = User::where("account_id", $account->id)->first();
 
     // Comprueba si el usuario está logueado
     $aux = false;
-    if ( Auth::user() && $user->id == Auth::user()->id ) {
+    if ( Auth::user() && Auth::user()->id == $user->id) {
       $aux = true;
     }
-
-    return view( "account.show", compact("user", "games", "categories", "aux") );
+    return view( "account.show", compact("account", "user", "aux") );
   }
 
   /**
-   * Show the form for editing the specified resource.
+   * Show the form for editing the specified user.
    *
    * @return \Illuminate\Http\Response
    */
   public function edit()
   {
     $user = Auth::user();
-    return view( "account.edit", compact("user") );
+    $account = Account::where("id", $user->account->id)->first();
+    return view( "account.edit", compact("user", "account") );
   }
 
   /**
-   * Update the specified resource in storage.
+   * Update the specified user in storage.
    *
    * @param  \Illuminate\Http\Request  $request
    * @param  int  $id
@@ -65,18 +54,27 @@ class AccountController extends Controller
   public function update(Request $request, $id)
   {
     $user = User::find($id);
-    $imgPath = $user->img;
-    $user->update( $request->all() );
+    $account = Account::where("id", $user->account->id)->first();
+
+    $imgPath = $account->img;
+    $user->update([
+      "name" => $request->name,
+    ]);
+    $account->update([
+      "slug" => Str::slug($request->name),
+      "desc" => $request->desc,
+    ]);
 
     if ( $request->file("img") ) {
-      Storage::disk("myDisk")->delete($imgPath);
-      $file = $request->file("img");
-      $path = Storage::disk("myDisk")->put("users", $file);
-      $user->img = $path;
+      if ( $imgPath != "users/default.webp" ) {
+        Storage::disk("myDisk")->delete($imgPath);
+      }
+      $path = Storage::disk("myDisk")->put("users", $request->file("img"));
+      $account->img = $path;
+      $account->save();
       $user->save();
     }
-
-     return redirect( action("AccountController@show", $user->name) );
+    return redirect( action("AccountController@show", $account->slug) );
   }
 
 }
